+++ src/components/publishedpost/atoms/PublishButton.stories.tsx
@@ -1,9 +1,9 @@
 import { Meta, Story } from '@storybook/react'
 import React, { ComponentProps } from 'react'
-import { StyledComponent as Component } from './DeleteButton'
+import { StyledComponent as Component } from './PublishButton'
 
 export default {
-  title: 'Delete Button',
+  title: 'Publish Button',
   component: Component
 } as Meta
 
@@ -19,7 +19,7 @@ const hoge = async (id: number) => {
 
 export const Default = Template.bind({})
 Default.args = {
-  className: "delete_button",
-  id: 2,
+  className: "publish_button",
+  id: '2',
   handleClickEvent: hoge
 }
