+++ src/lib/types/PostProps.ts
@@ -1,0 +1,10 @@
+export type PostProps = {
+  id: string
+  title?: string
+  author?: {
+    name: string
+    email: string
+  } | null
+  content?: string
+  published?: string
+}
