+++ src/pages/api/publish/[id].ts
@@ -1,6 +1,11 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import prisma from "../../../lib/prisma";
 
+/**
+ * PUT /api/publish/:id
+ * @param req request
+ * @param res response
+ */
 export default async function handle(req: NextApiRequest, res: NextApiResponse): Promise<void> {
   // TODO:find better way
   const postId = typeof req.query.id === 'string' ? req.query.id : undefined
