--- src/pages/form.tsx
@@ -1,8 +1,0 @@
-import { NextPage } from "next";
-import { SimpleForm } from "../components/simpleForm";
-
-const Form: NextPage = () => (
-  <SimpleForm />
-)
-
-export default Form
